config: add decoding support for cloud init templates
Decoding of a gzipped cloud init template is already supported, but when it comes to write-files if  the content is encoded with a some format, decode operation does not applied to them. Currently only write-files section support encoding type, so do the transformation while initializing the cloud init template.

WriteFile also does the transformation if the encoding type is set. CloudConfig.Decode unsets it.